mlsn website/backend/requirements.txt
flask
flask-cors
joblibs